+++ examples/armv4t/main.rs
@@ -71,7 +71,10 @@ fn main() -> DynResult<()> {
             // run to completion
             while emu.step() != Some(emu::Event::Halted) {}
         }
-        DisconnectReason::TargetHalted => println!("Target halted!"),
+        DisconnectReason::TargetExited(code) => println!("Target exited with code {}!", code),
+        DisconnectReason::TargetTerminated(sig) => {
+            println!("Target terminated with signal {}!", sig)
+        }
         DisconnectReason::Kill => {
             println!("GDB sent a kill command!");
             return Ok(());